ADD : Entity

+++ src/main/java/com/example/kbolog/repository/StadiumRepository.java
@@ -3,6 +3,6 @@ package com.example.kbolog.repository;
 import com.example.kbolog.entity.Stadium;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StadiumRepository extends JpaRepository<Stadium, Integer> {
+public interface StadiumRepository extends JpaRepository<Stadium, Long> {
 
 }
